+++ packages/veritone-widgets/src/redux/configureStore.development.js
@@ -7,7 +7,7 @@ import {
   getBaseStoreEnhancers,
   getDevOnlyStoreEnhancers
 } from './store';
-import createRootReducer from './reducer';
+import createRootReducer from './rootReducer';
 import rootSaga from './rootSaga';
 
 // Redux devtools browser extension hook
