+++ packages/veritone-widgets/src/redux/rootReducer.js
